requirements.txt
pandas
selenium
webdriver-manager
python-dotenv
beautifulsoup4
pdfplumber
plotly
requests